Automate cache (#2)
* Adding some automation for uploading cache 2 db

* Moving things around

* Refactored CLI out into its own file to allow package to be used as a module.

* Removing my code, putting it in another project

* Cleanup for PR
