+++ src/index.js
@@ -1,0 +1,6 @@
+import dotenv from 'dotenv';
+dotenv.config();
+
+export * from './on-chain/index.js';
+export * from './output.js';
+export * from './utils.js';
